Change the PUT URL syntax and add missing existence check.
The PUT URL was changed to agree with clean URL conventions. You can
now strip off the query portion and use it as a GET URL to retrieve
the resource. There was also a missing existence check that a client
would use to save bandwith by not having to upload a c-block that the
storage node was already storing.