Break GET handling into seperate file and change URL syntax.
The GET URL syntax was changed to agree with clean URL conventions.
The client can now specify a list of c-indexes to skip. It can also request a
particular c-index, these requests can be pipelined of course. This gives
the client the ability to list all the c-blocks it wants or doesn't want.
It now returns the X-Block-N headers before the c-blocks so that the client
can act in response to the data as it's being received. However this adds a
a delay between the generation of the X-Block-N headers and the reading
of the files in which they can disappear. It also uses potentially unbounded
memory, but we expect that the erasure coding scheme will limit the
number of possible c-blocks to 255.

P.S. Actually it always used potentially unbounded memory because the
footers were buffered by libmicrohttpd anyway.